Lazy-load ETF spider + price/returns/CAGR charts (parity with stocks)
Apply the stock-side two-layer chart-deferral pattern (dynamic({ssr:
false}) + useInView) to the ETF detail page's two heaviest chart
surfaces:

1. EtfRadarChart (spider) — already had ssr:false, now gated on viewport
   intersection so the chart.js + react-chartjs-2 chunk fetch + eval no
   longer lands in the hydration burst.

2. EtfChartTabs (price / returns / CAGR) — previously imported
   PriceChart, chart.js, and react-chartjs-2 at module top, which shipped
   chart.js in the main client bundle on every ETF detail page render.
   Split into:
     - EtfChartTabs.tsx (shell): tab UI, range pill, tab state — no
       chart.js imports
     - EtfChartTabsChartBody.tsx (heavy): chart.js register, PriceChart,
       react-chartjs-2 Bar, PerformanceBars — dynamically imported by
       the shell with ssr:false + useInView gate

PRICE_CHART_RANGES is inlined in the shell instead of imported from
PriceChart to avoid side-effect-loading chart.js for a 5-element
constant. Skeletons match the live chart heights (h-64 sm:h-72) so
swapping in the real chart causes zero layout shift.

Mirrors PriceChartLazy.tsx / QuarterlyMetricsChartLazy.tsx /
TickerRadarChart.tsx on the stock side — no stock files touched.
